# frontend/openapi/asbl-club-api.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "asbl.club API",
    "description" : "Public read API exposing associations and their public events. Reads are open with CORS for showcase sites. Everything else requires a JWT access token from POST /api/v1/auth/login, sent as a Bearer token.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/v1/asbls/{slug}/manage/events/{eventId}" : {
      "get" : {
        "tags" : [ "Event management" ],
        "summary" : "One event with its ticket categories and sales (members)",
        "operationId" : "getManagedEvent",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Event management" ],
        "summary" : "Edit a draft or published event (administrators)",
        "operationId" : "updateEvent",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The updated event",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          },
          "409" : {
            "description" : "Cancelled or past event (EVENT_NOT_EDITABLE), or a shorter cancellation delay once tickets are sold (CANCELLATION_DAYS_LOCKED)",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Event management" ],
        "summary" : "Delete a draft event (administrators)",
        "operationId" : "deleteDraftEvent",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "409" : {
            "description" : "Not a draft (EVENT_NOT_EDITABLE): cancel a published event instead",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
    "/api/v1/asbls/{slug}/manage/events/{eventId}/tickets/{ticketId}" : {
      "put" : {
        "tags" : [ "Event management" ],
        "summary" : "Change a ticket category (administrators)",
        "operationId" : "updateTicketCategory",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "ticketId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddTicketRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The updated event",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          },
          "409" : {
            "description" : "Fewer seats than taken (SEATS_BELOW_SOLD), or event not editable",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Event management" ],
        "summary" : "Remove a ticket category nobody booked (administrators)",
        "operationId" : "removeTicketCategory",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "ticketId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The updated event",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          },
          "409" : {
            "description" : "Already booked (TICKET_IN_USE), or event not editable",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/asbls" : {
      "post" : {
        "tags" : [ "Associations" ],
        "summary" : "Create an association; I become its administrator",
        "operationId" : "createAsbl",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAsblRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AsblResource"
                }
              }
            }
          },
          "409" : {
            "description" : "URL identifier or BCE number already taken",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/asbls/{slug}/manage/events" : {
      "get" : {
        "tags" : [ "Event management" ],
        "summary" : "All the association's events, drafts included (members)",
        "operationId" : "listManagedEvents",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEventList"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Event management" ],
        "summary" : "Create an event, as a draft (administrators)",
        "operationId" : "createEvent",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
    "/api/v1/asbls/{slug}/manage/events/{eventId}/tickets" : {
      "post" : {
        "tags" : [ "Event management" ],
        "summary" : "Add a ticket category to an event (administrators)",
        "operationId" : "addTicketCategory",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddTicketRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Added; the updated event",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/asbls/{slug}/manage/events/{eventId}/publish" : {
      "post" : {
        "tags" : [ "Event management" ],
        "summary" : "Publish a draft event (administrators)",
        "operationId" : "publishEvent",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/asbls/{slug}/manage/events/{eventId}/cancel" : {
      "post" : {
        "tags" : [ "Event management" ],
        "summary" : "Cancel a published event: hidden and no longer bookable (administrators)",
        "operationId" : "cancelEvent",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The cancelled event",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManagedEvent"
                }
              }
            }
          },
          "409" : {
            "description" : "Not a published event (EVENT_NOT_EDITABLE)",
            "content" : {
              "application/problem+json" : { }
            }
          }
        },
        "security" : [ {
          "bearer" : [ ]
        } ]
      }
    },
...
    "/api/v1/events/{eventId}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get a public event with its association, location and tickets",
        "operationId" : "getEvent",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicEvent"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/asbls/{slug}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get an association by slug",
        "operationId" : "getAsbl",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AsblResource"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateEventRequest" : {
        "required" : [ "startsAt", "title", "visibility" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "visibility" : {
            "minLength" : 1,
            "pattern" : "PUBLIC|MEMBERS",
            "type" : "string"
          },
          "cancellationDays" : {
            "maximum" : 365,
            "minimum" : 0,
            "type" : "integer",
            "description" : "Until how many days before the start buyers may cancel their ticket and be refunded; 0: not refundable. Optional: 0 when creating, unchanged when editing. Once a ticket is sold, it can only grow.",
            "format" : "int32"
          }
        }
      },
      "ManagedEvent" : {
        "required" : [ "canManage", "canSeeAttendees", "cancellationDays", "id", "startsAt", "status", "tickets", "title", "visibility" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "PUBLISHED", "CANCELLED", "ENDED" ]
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "MEMBERS" ]
          },
          "cancellationDays" : {
            "type" : "integer",
            "description" : "Until how many days before the start buyers may cancel; 0: never",
            "format" : "int32"
          },
          "canManage" : {
            "type" : "boolean"
          },
          "canSeeAttendees" : {
            "type" : "boolean",
            "description" : "Administrators and treasurers see who booked"
          },
          "tickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ManagedTicket"
            }
          }
        }
      },
      "ManagedTicket" : {
        "required" : [ "id", "label", "price", "soldSeats", "totalSeats" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "label" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number",
            "description" : "Price in euros"
          },
          "totalSeats" : {
            "type" : "integer",
            "format" : "int32"
          },
          "soldSeats" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "AddTicketRequest" : {
        "required" : [ "label", "price", "totalSeats" ],
        "type" : "object",
        "properties" : {
          "label" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "price" : {
            "minimum" : 0.0,
            "exclusiveMinimum" : false,
            "type" : "number"
          },
          "totalSeats" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "CreateAsblRequest" : {
        "required" : [ "bceNumber", "defaultLanguage", "denomination", "slug" ],
        "type" : "object",
        "properties" : {
          "denomination" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "bceNumber" : {
            "minLength" : 1,
            "pattern" : "\\d{4}\\.\\d{3}\\.\\d{3}",
            "type" : "string"
          },
          "slug" : {
            "maxLength" : 255,
            "minLength" : 0,
            "pattern" : "[a-z0-9-]+",
            "type" : "string"
          },
          "defaultLanguage" : {
            "minLength" : 1,
            "pattern" : "fr|nl|en",
            "type" : "string"
          }
        }
      },
      "AsblResource" : {
        "required" : [ "denomination", "slug" ],
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "denomination" : {
            "type" : "string"
          }
        }
      },
...
      "PublicEvent" : {
        "required" : [ "asbl", "id", "startsAt", "tickets", "title" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "type" : "string"
          },
          "asbl" : {
            "$ref" : "#/components/schemas/AsblResource"
          },
          "cancellableUntil" : {
            "type" : "string",
            "description" : "Until when buyers may cancel their ticket and be refunded; absent: not refundable on request",
            "format" : "date-time"
          },
          "tickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PublicTicket"
            }
          }
        }
      },
...
      "ManagedEventList" : {
        "required" : [ "canManage", "denomination", "events", "slug" ],
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "denomination" : {
            "type" : "string"
          },
          "canManage" : {
            "type" : "boolean",
            "description" : "Whether the viewer may create and change events"
          },
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ManagedEventItem"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}